// src/database/generated/prisma/sql/insertSchoolToDatabase.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck
import * as $runtime from "@prisma/client/runtime/client";

/**
 * @param text
 * @param text
 * @param float8
 * @param float8
 */
export const insertSchoolToDatabase = $runtime.makeTypedQueryFactory('INSERT INTO "School" (id, name, address, longitude, latitude)\nVALUES (gen_random_uuid(), $1, $2, $3, $4) RETURNING *;') as (text: string, text: string, float8: number, float8: number) => $runtime.TypedSql<insertSchoolToDatabase.Parameters, insertSchoolToDatabase.Result>;

export namespace insertSchoolToDatabase {
    export type Parameters = [text: string, text: string, float8: number, float8: number];
    export type Result = {
        id: string;
        name: string;
        address: string;
        latitude: number;
        longitude: number;
    };
}